Fix crash if report response is null

+++ src/Listener/SendPushNotifications.php
@@ -110,9 +110,9 @@ class SendPushNotifications
          * @var MessageSentReport $report
          */
         foreach ($webPush->flush() as $report) {
-            if (!$report->isSuccess() && in_array($report->getResponse()->getStatusCode(), [401, 403, 404, 410])) {
+            if (!$report->isSuccess() && $report->getResponse() && in_array($report->getResponse()->getStatusCode(), [401, 403, 404, 410])) {
                 PushSubscription::where('endpoint', $report->getEndpoint())->delete();
-            } elseif (!$report->isSuccess()) {
+            } elseif (!$report->isSuccess() && $report->getResponse()) {
                 echo "[x] Message failed to sent for subscription {$report->getEndpoint()}: {$report->getReason()}";
             }
         }
